Preprocess single project not corpus of repos
Having a look through the repository, I could not find a way to input a
single local Python project to the preprocessor (which is the first
stage of DLTPy), only a list of GitHub repositories. Unfortunately, the
code is not modular enough to just add a small frontend, so I had to
rewrite a bit. I actually replaced the repo-logic with a local directory
one instead of adding to it. I might regret that later but I think I can
just re-add it then.
Apart from that, automatic formatting was applied in some files and some
warnings which were emitted by the linter were fixed.