Update dependencies to latest
Just a blatant and careless upgrade of all dependencies since I could
not bear it any longer to be required to run this code with Python 3.7
(instead of 3.8 or 3.9) but also experience a ton of non-obvious
library-internal breakages which I could patch in some instances by
overwriting methods or installing/forcing extra packages. It started to
amass.